backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
celery==5.4.0
redis==5.1.1
PyGithub==2.4.0
google-generativeai==0.8.3
chromadb==0.5.15
sentence-transformers==3.1.1
pathspec==0.12.1
python-dotenv==1.0.1
httpx==0.27.2
